Refactoring
Changed primaryButton.jsx to primary-button.jsx to comply with more standard naming syntax.
Renamed and then moved welceom-page.jsx into pages folder for better organization.
