+++ src/App.jsx
@@ -1,4 +1,4 @@
-import Welcome from './components/welcome';
+import Welcome from './pages/welcome-page';
 
 export default function App() {
   return (
