+++ src/main.jsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App.jsx';
 import './styles.css';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
+import App from './App.jsx';
+import About from './components/about.jsx';
 
 const router = createBrowserRouter([
   {
